pyproject.toml
[project]
name = "sre-copilot"
version = "0.1.0"
description = "SRE Copilot - RAG-based documentation assistant with MCP servers"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "SRE Team" }]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "chromadb>=0.4.0",
    "httpx>=0.25.0",
    "structlog>=23.0.0",
    "click>=8.0",
    "rich>=13.0",
    "mcp>=1.0.0",
    "tenacity>=8.0",
    "python-frontmatter>=1.0.0",
    "tiktoken>=0.5.0",
]

[project.optional-dependencies]
watsonx = [
    "langchain-ibm>=0.1.0",
    "ibm-watsonx-ai>=0.2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
sre-copilot = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
